warnings
git-svn-id: svn://localhost/trunk@983 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/surfit/hist.cpp
@@ -83,7 +83,7 @@ size_t d_hist::operator()(REAL value) const
 		return UINT_MAX;
 	}
 
-	size_t pos = (value-z_from)/z_step;
+	size_t pos = (size_t)((value-z_from)/z_step);
 	return MIN(size()-1,pos);
 };
